pyproject.toml
[project]
name = "jdbc-mcp-server"
version = "0.1.0"
description = "MCP server for database connectivity supporting PostgreSQL, MySQL, SQLite, and DB2"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}

dependencies = [
    "fastmcp>=2.0.0",
    "pydantic>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "mysql-connector-python>=8.0.0",
    "ibm_db>=3.2.0",
    "sqlparse>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["setuptools>=65.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["jdbc_mcp_server", "jdbc_mcp_server.database"]
package-dir = {"" = "src"}

[tool.black]
line-length = 100
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
